🚚 zv: internal: Move crate-specific impls to super module
This will come in handy in a following commit, where we move most of the
parsed module to `zvariant_utils`.
